requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
install==1.3.5
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
tzdata==2024.1
yarl==1.9.4
